Update WebApplicationExtensions for Bootstrap
