+++ .gitignore
@@ -26,6 +26,7 @@ publish/
 debugging/
 deprecated/
 artifacts/
+App_Data/
 
 # Visual Studio 2015 cache/options directory
 .vs/
